Initial work on file processor
This adds a new file processor mode. So the command changes from:

youtube-cue https://youtu.be/watch?v=dfjosufoadsi

to additionally supporting:

youtube-cue description.nfo

This however results in some problems. It is hard to get album and album
artist information from a plain text file, so we instead now have 2
additional flags:

--cue-performer "Performer for the entire track collection"
--cue-title "Title for the entire collection"

They usually, (but not always) will correspond to the album artist and
album title.

The changes are fairly obvious - I've also tried to make this more
modular, so hopefully this is easier to use as a module.